src-django/requirements.txt
asgiref==3.8.1
build==1.2.2.post1
Django==5.1.6
djangorestframework==3.15.2
packaging==24.2
pyproject_hooks==1.2.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
